+++ web/src/components/ZTag/index.js
@@ -1,0 +1,4 @@
+import ZTag from './ZTag';
+
+export { ZTag };
+export default ZTag;